+++ applications/feeds/urls.py
@@ -1,13 +1,7 @@
 from rest_framework.routers import DefaultRouter
-from feeds.views import (
-    EntryViewSet,
-    BulkEntryView,
-    LocationViewSet,
-    AreaViewSet,
-    AreasCountViewSet,
-    CityByCityCountView,
-    AreaLiteViewSet,
-)
+from feeds.views.areas import AreaViewSet, AreaLiteViewSet, AreasCountViewSet, CityByCityCountView
+from feeds.views.locations import LocationViewSet
+from feeds.views.entries import EntryViewSet, BulkEntryView
 from django.urls import path
 
 router = DefaultRouter(trailing_slash=False)
@@ -15,7 +9,7 @@ router.register("locations", LocationViewSet, basename="list-locations")
 router.register("entries", EntryViewSet, basename="create-entry")
 router.register("areas", AreaViewSet, basename="list-area")
 router.register("areas/count", AreasCountViewSet, basename="count-area")
-router.register("areas-lite", AreaLiteViewSet, basename="list-arealite")
+router.register("areas-lite", AreaLiteViewSet, basename="list-area-lite")
 
 urlpatterns = [
     path("entries/bulk", BulkEntryView.as_view()),
